Don't flush only if file handle is closed
This matches directly the comment. Previous implemetation
was not flushing when Input() was different than cli_CurrentInput
which is the case when running script. This caused the arguments
of the last command in script to be re-emited as another command.
This was not visible for shell commands as they ReadArg the Input()